refactor: Update return statement in DELETE function
The return statement in the DELETE function has been refactored. Previously, it was using NextResponse.json method with null as the first argument and an object containing status code 204 as the second argument. Now, it's directly creating a new instance of NextResponse with null as the first argument and an object containing status code 204 as the second argument. This change simplifies the return statement and makes it more readable.