refactor: Simplify type name generation in scripts
The code has been refactored to simplify the generation of type names in various script files. A new constant, `typeName`, has been introduced to hold the capitalized version of the key, which is then used throughout the rest of the script. This change reduces redundancy and improves readability by avoiding repeated string manipulation operations.