refactor: Simplify POST method in generate-route-content
The POST method in the generate-route-content script has been refactored for simplicity. The previous implementation checked if an item already existed in the database and updated it, or created a new one if it didn't exist. This logic has been removed, and now all incoming items are simply added to the database. The status code returned is always 201, indicating successful creation of a resource.