feat: Add unique ID to new items in POST request
The code now generates a unique ID for each new item created through the POST request. This is done by spreading the request body into a new object and adding an 'id' property with a randomly generated UUID. The updated item is then pushed to the database and written to file.