feat: Add filtering functionality to data retrieval
Implemented a new feature that allows for the application of filters when retrieving data. This was achieved by creating an 'operators' object and an 'applyFilters' function, which applies the specified filters to the data before it is returned. The filtered data is then paginated and returned in the GET request response.