feat: Add inner join functionality to route content generator
A new function `applyInnerJoin` has been added to the route content generation script. This function allows for joining related data from a separate resource based on a common key, enhancing the richness of the returned data. The GET method now uses this function to include related 'comments' in its response.