feat: Setup Postgres database and migrations
Adds Postgres as the database.
Configures Drizzle ORM for migrations.
Sets up scripts for generating and running migrations.
Adds a script to automatically generate a schema based on db.json.
Adds a script to initialize the database and run migrations on startup.
Adds concurrently and nodemon to watch DB changes for local development